docs: atomfeat: set up basic documentation for AtomFeat
Set up basic documentation for the AtomFeat type that generates autodocs
for the structure, its constructors, and interfaces.

Edit the existing documentation within code a little, and ensure type
safety for `logspaced` everywhere.

Note:
The primary inner constructor for AtomFeat is yet to be documented.
When the signature was included in the atomfeat.md file, Documenter gave
a Missing Docstring warning.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>