docs: atomgraph: set up basic documentation for AtomGraph
Set up basic documentation for the AtomFeat type that generates autodocs
for the structure, its constructors, and interfaces.

Note:
Documenter raises an error -
	ERROR: LoadError: UndefVarError: SimpleWeightedGraph not defined
if AtomGraph's either constructor signature is mentioned under autodocs.
For this reason, and since the purpose of a constructor is fairly obvious,
plain markdown has been used to mention the constructors' signatures instead.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>