src: pmg_graphs: prevent warnings from being generated for folders
If `input_folder` houses directories in it, a warning gets thrown saying
the graph couldn't be built for that directory, which given the context,
shouldn't be happening in the first place as nobody shouldn't even be
trying to construct a graph from an 'input file' that is a directory.

So, instead, update file_list to actually be a list of files
ONLY, and not include directories too.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>