src: pmg_graphs: add option for overwrites
Addresses #45.

Add option for overwriting in `build_graphs_batch()` (default value of
overwrite = true).

If `overwrite = false`, first check if the output directory exists,
and create a new list (`existing_list`) that contains names of all the files
in the output directory. While building the graphs, cross-reference with
`existing_list` to  check if there already exists a file with the same
name (which means this graph exists and was previously built), and continue
execution accordingly.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>