test: atomgraph_tests: add test for overwrite parameter
Commit a441cf7 introduced a new parameter for `build_graph_batches()`
called `overwrite`.

Create two test cases that test this parameter. The idea behind both the
testcases is the same. Call `build_graph_batches()` and check if the
last modified time for the files in the `output_folder` changes or not.

Cleanup afterwards, and remove `output_folder` and its contents after
testing.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>