src: pmg_graphs: fix overwrite condition
Consider the case where overwrite=false, and output_directory doesn't
exist. In this case, !overwrite is true, and isdir(output_directory) is
true as well (since we create an output_directory if it doesn't exist
already).
However, the check (id in existing_files) gives an error as existing_files
isn't defined at this point (existing_files is defined only when !overwrite
and isdir(output_directory)). So, an undefined variable is illegally
accessed.
Fix this issue by increasing the scope of existing_files (by one level)
and updating the check to see if there are existing_files that will need
to be skipped, and not overwritten.

Fixes: a441cf7 ("src: pmg_graphs: add option for overwrites")
Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>