src: pmg_graphs: fix return value add new test
Consider the case, where build_graphs_batches() is called and there
already exists an output_folder, and overwrite = false.

In this case, the build_graph() function isn't called (as expected), but
the existing graphs aren't returned to the caller; which is not the
expected behaviour.

Also, add a test which checks for this condition.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>