fix tests, and other regressions introduced in 7c8a495
* Rename all previous mentions of `AtomFeatureDescriptor` appropriately
(to `ElementFeatureDescriptor` or `SpeciesFeatureDescriptor`)

* Update instances where `.Utils.AtomFeatureUtils` module is used to
`.Utils.ElementFeatureUtils`.

* `encode_f` and `decode_f` were removed as fields of
`ElementFeatureDescriptor`. In their place, define the actual
encode/decode functions directly in the interface which used to be
responsible for calling them anyways.
This is reasonable for `ElementFeatureDescriptor`, since encode/decode
functions for it will always be the same, and depend only on the
parameters that are stored.

* Add an `nbins` field to `ElementFeatureDescriptor`. Since `encode_f`,
and `decode_f` are no longer fields of `ElementFeatureDescriptor`, there
is no real way to encode features into the bins as required by the user.
Having `nbins` as a field allows us to do this, and also makes sense
intuitively as well (since `ElementFeatureDescriptor` essentially
describes the encoding/decoding procedure and binning is a part of that).

* Remove the custom "silly encode/decode" tests for
`ElementFeatureDescriptor`. These could however be later re-introduced
in some other manner for testing `SpeciesFeatureDescriptor`.

* Fix syntax error caused by a missing closing parentheses in
`decode(::AtomGraph)`

Fixes: 7c8a495 ("new type hierarchy for features, not done fixing things
so tests will fail. Still to do, at a minimum: fix tests, plus decide on
terminology (encoder probably?)")
Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>