elementfeature.jl: integrate `EncoderDecoder` with `ElementFD`
Replace `nbins` and `logspaced` fields in `ElementFeatureDescriptor`
with `encoder_decoder` which is of type `EncoderDecoder`.

The default for `encoder_decoder` this is `DummyED` ("dummy" as defined in
statistical literature). Accordingly, the default encoding and decoding
functions for `DummyED` are `default_efd_encode`, and
`default_efd_decode` respectively, which have also been defined.

The interface of `ElementFeatureDescriptor` itself remains largely
unchanged, since only the implementation internals have been changed.

A new enum type EncodeOrDecode has also been created. This specifies
two values - ENCODE and DECODE. This is used (typically internally by a
FeatureDescriptor only) when a call for encoding/decoding is made to the
callable `EncoderDecoder` subtype.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>