fix `decode()` dispatches for `FeatureDescriptors` and `Featurizations`
`ElementFeatureDescriptor` and `GraphNodeFeaturization` both define
their own implementations of `decode()`. This causes a lot of problems
(due to two different modules having a similar name), and results in
tests failing tests. Fix this issue in a convenient way, similar to what
was done with `encodable_elements()` in 58a3e8f.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>