featurizations: rename the `featurize` function to `encode`
* Rename the `featurize` function to `encode`.
Since `FeaturizedAtoms` was introduced, and `encoded_features`,
`featurization` were removed as fields in `AtomGraph`,
`featurize!` was renamed to `featurize` in 043878d.
However, since this function no longer "featurizes" (i.e., it doesn't
attach any encoded features to an `Atoms` object), and returns a set of
`encoded_features` instead, it would be better, and more intuitive to
rename this as `encode`.

* For symmetry with existing implementations of `encode`, and also their
converse, `decode` functions, reverse the order of arguments to be
`(featurization, object_to_be_encoded)`.

* Update all instances of `featurize` to`encode`.

* Also, export a generic `encode` from the `Featurization` module
itself, which will throw an error if the `encode` function for a
given combination of argument types isn't defined.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>