features: simplify encoding-decoding logic, and minimize repetitive code
* Minimize repetitive code that would be required to be written while
implementing `encode` and `decode` functionality. All
`FeatureDescriptor`s must have a `Codec` attached. The responsibility of
actually performing the encoding/decoding is delegated to the
`encode_f`/`decode_f` functions of the attached `Codec` respectively.
This would mean that, the definitions for all
`encode(::AbstractFeatureDescriptor, ::AbstractAtoms)` and
`decode(::AbstractFeatureDescriptor, encoded_feature)` functions for all
feature descriptors would look the same.
Also, all `Codec`s must be callable and define an implementation taking
whatever `FeatureDescriptor` they support as the first argument, and
an `AbstractAtoms` object (in case of encoding) or an `encoded_feature`
(in case of decoding) as the second argument (since `encoded_feature`
will no longer be a subtype of `AbstractAtoms`, as of 99618bc).
So, instead of throwing the at the level of the generic `encode`/`decode`,
throw an error (an `ErrorException`) at the level of the generic
callable `Codec` object's syntax, when an invalid combination of
arguments (for which no implementation is present) is executed.

* Modify the test checking for a `MethodError` while performing `decode`
(for `FeatureDescriptor`s) to check for an `ErrorException` instead.

* Remove the `EncodeOrDecode` enum, as it doesn't really serve any real
purpose anymore. The work done to simplify the encoding-decoding logic
in itself renders it unnecessary.
However, regardless of this, `EncodeOrDecode` isn't required anymore.
The `EncodeOrDecode` enum was introduced back when `Atom`s objects used
to be mutated to represent the encoded features as well. However, as of
99618bc, this is no longer the case, and `Atom`s objects are no longer
as tightly coupled to encoded features.
As a result of this, even if an `encoded_features` is essentially an
`Atom`s object that has its values modified, it won't be of a subtype
`AbstractAtom`.
This further means that, all calls being made for `decode` operations
using the callable syntax of a `Codec`, like (`OneHotOneCold`) will be
dispatched using only
`(ed::OneHotOneCold)(efd::ElementFeatureDescriptor, encoded_feature)`, and
`(ed::OneHotOneCold)(efd::ElementFeatureDescriptor, a::AbstractAtoms)`
would be dispatched only for `encode` operations; no longer for
`encode` and possibly `decode` too. This renders the `EncodeOrDecode`
enum unnecessary, as it is not really of any use.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>