orbitalfeature: annotate, fix order of shells in lookup in defaultdecode
* Add comments, and annotate `default_ofd_decode()`

* Turns out, instead of having the electronic configuration for an
element be represented in the order shells are actually filled, the
lookup_table stores them in a more "readable" way, i.e., numerically
sorted way.
For instance, the valence shell configuration for "Fe" is `4s2.3d6` if
you consider the order in which electrons are filled in the shells, but
in the default lookup_table, the order is `3d6.4s2`.
So, sort out the `shell_conf` first, before looking it up.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>