docs: fix doctests
Fix doctests and add a doctest that checks the return type of `decode()`
on a `FeaturizedAtoms` object. This is necessary because in the doctest
that tests for value, the value was getting truncated and hence was
`collect()`-ed. The return type isn't checked in this case because of
this.
Hence, explicitly check the `typeof()` also.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>